pyproject.toml
[project]
name = "sphinx-gooey"
version = "0.1.0"
description = "Produces a gallery of examples for Sphinx documentation. Totally language agnostic, include examples in any programming language in your documentation."
authors = [{ name = "Bryan Weber", email = "bryan.w.weber@gmail.com" }]
dependencies = ["sphinx-design>=0.2.0", "myst-parser>=0.18.0"]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "BSD-3-Clause" }

[project.optional-dependencies]
notebook = [
    "myst-nb>=0.16.0",
    "nbformat>=5.7.0",
    "jupytext>=1.14.7",
]

[project.entry-points."sphinx_gooey.generators"]
jupyter = "sphinx_gooey.jupyter:md_generator"
python = "sphinx_gooey.python:md_generator"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.isort]
profile = "black"

[tool.mypy]
check_untyped_defs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "sphinx.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "myst_nb.*"
ignore_missing_imports = true

[tool.pylsp-mypy]
enabled = true
overrides = ["--python-executable", ".venv/bin/python"]

[tool.ruff]
target-version = "py311"

[tool.pdm.scripts]
doc = { cmd = [
    "sphinx-build",
    "-b",
    "html",
    "-W",
    "--keep-going",
    "doc",
    "doc/_build/html",
    "-a",
] }
livedoc = { cmd = [
    "sphinx-autobuild",
    "doc",
    "doc/_build/html",
    "--ignore",
    "*examples*",
    "--watch",
    "src",
    "--port",
    "0",
    "--open-browser",
] }
test = { cmd = ["pytest", "--cov=src/sphinx_gooey", "tests"] }

[tool.pdm.dev-dependencies]
doc = [
    "sphinx>=5.0.2",
    "sphinx-autobuild>=2021.3.14",
    "myst-parser>=0.18.0",
    "sphinx-design>=0.2.0",
    "pydata-sphinx-theme>=0.9.0",
    "myst-nb>=0.16.0",
]
dev = [
    "mypy>=0.982",
    "black>=22.10.0",
    "pre-commit>=2.20.0",
    "isort>=5.10.1",
    "ruff>=0.0.277",
]
test = ["pytest>=7.4.0", "pytest-cov>=4.1.0"]
mypy = ["types-docutils>=0.20.0.1"]
